ai-backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
grpcio==1.62.2
grpcio-tools==1.62.2
protobuf==4.25.3
sentence-transformers==2.6.1
faiss-cpu==1.8.0
openai==0.28.1
numpy>=1.21.0,<2.0.0
python-dotenv==1.0.1 
